sphinx-agent: Move `Uri` implementation into its own file
